requirements.txt
playwright==1.54.0
PyYAML==6.0.2